+++ src/api/giphy/types.ts
